requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0
Werkzeug==2.0.1
Jinja2==3.0.1
SQLAlchemy==1.4.23
email-validator==1.1.3
pymysql==1.0.2